Some random additions
Made three scenes for food items; good, neutral, and toxic. They are all the same with only a placeholder sprite and no functionality at all. Removed some code that shouldn't be used. Started a little camera controller function to move the camera slightly in the direction the player is moving for better visibility.