First Release
Still quite buggy. The pause menu buttons don't respond to clicking for an unknown reason. There is no kill plane when falling off the stage. The UI bounces when the camera is against the left stage limit. The camera gets stuck in place when it reaches the right limit. The player can push the camera to any camera boundary, and continue off-screen. The top and bottom camera boundaries do not work for an unknown reason. There are no sounds effects implemented despite being in the game files. The bushes do not work (intended behavior is to be able to press 'E' when next to them to receive food and regain stamina). Only renders at 1080p regardless of monitor resolution.